Tic Tac PRO
This is the ultimate Tic Tac Toe project.

I used the miniMax algorithm to create an AI that never loses, yes never! :)
It works by checking all the possible solutions for the board at a certain time, and assigns a value to each try.
If it is the maximizing players  turn, in this case the "X", it will return the possibility that return the higgest possible value, while on the other hand if it is the minimizing players turn it will choose the path that leads to the smallest value.

I also used the Firebase real-time database to allow people to play online against each other.